pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "confluence-markdown-exporter"
dynamic = ["version"]
description = "A tool to export Confluence pages to Markdown"
keywords = ["confluence", "atlassian", "markdown", "export", "convertion", "download"]
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Sebastian Penhouet" }
]
requires-python = ">= 3.10"
dependencies = [
    'atlassian-python-api',
    'jmespath',
    'markdownify',
    'pandas',
    'pydantic-settings',
    'pyyaml',
    'questionary',
    'tabulate',
    'tqdm',
    'typer',
]

[project.optional-dependencies]
dev = [
    'ruff==0.9.10',
    'pipx'
]

[project.urls]
Homepage = "https://github.com/Spenhouet/confluence-markdown-exporter"
Documentation = "https://github.com/Spenhouet/confluence-markdown-exporter/blob/main/README.md"
Source = "https://github.com/Spenhouet/confluence-markdown-exporter"
Tracker = "https://github.com/Spenhouet/confluence-markdown-exporter/issues"

[project.scripts]
confluence-markdown-exporter = "confluence_markdown_exporter.main:app"
cf-export = "confluence_markdown_exporter.main:app"

[project.entry-points."pipx.run"]
confluence-markdown-exporter = "confluence_markdown_exporter.main:app"
cf-export = "confluence_markdown_exporter.main:app"

[tool.setuptools]
packages = ["confluence_markdown_exporter", "clients"]

[tool.setuptools_scm]